+++ packages/ui/src/index.ts
@@ -1,10 +1,9 @@
-import { DirectionProvider } from '@base-ui/react/direction-provider';
-
 export * from './badge';
 export * from './button';
 export * as Card from './card';
 export * as Collapsible from './collapsible';
 export * as CollapsibleCard from './collapsible-card';
+export * from './direction-provider';
 export * as AlertDialog from './alert-dialog';
 export * as Dialog from './dialog';
 export * as Drawer from './drawer';
@@ -25,5 +24,3 @@ export * as Tooltip from './tooltip';
 export { getWpCompatOverlaySlot } from './utils/wp-compat-overlay-slot';
 export { useEnableWpCompatOverlaySlot } from './utils/use-enable-wp-compat-overlay-slot';
 export * from './visually-hidden';
-
-export { DirectionProvider };